refactor: extract lifecycle and session hooks
