+++ src/hooks/useSessionManager.ts
@@ -1,0 +1,238 @@
+import { useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { useConnections } from '../contexts/ConnectionContext';
+import { Connection, ConnectionSession } from '../types/connection';
+import { SettingsManager } from '../utils/settingsManager';
+import { StatusChecker } from '../utils/statusChecker';
+import { ScriptEngine } from '../utils/scriptEngine';
+
+export const useSessionManager = () => {
+  const { t } = useTranslation();
+  const { state, dispatch } = useConnections();
+
+  const settingsManager = SettingsManager.getInstance();
+  const statusChecker = StatusChecker.getInstance();
+  const scriptEngine = ScriptEngine.getInstance();
+
+  const [activeSessionId, setActiveSessionId] = useState<string | undefined>();
+
+  const connectSession = async (session: ConnectionSession, connection: Connection) => {
+    const settings = settingsManager.getSettings();
+    const startTime = Date.now();
+
+    settingsManager.logAction('info', 'Connection initiated', connection.id, `Connecting to ${connection.hostname}:${connection.port}`);
+
+    try {
+      await scriptEngine.executeScriptsForTrigger('onConnect', { connection, session });
+    } catch (error) {
+      console.error('Script execution failed:', error);
+    }
+
+    if (connection.statusCheck?.enabled) {
+      statusChecker.startChecking(connection);
+    }
+
+    const timeout = (connection.timeout || settings.connectionTimeout) * 1000;
+    const connectionPromise = new Promise<void>((resolve) => {
+      setTimeout(() => {
+        const connectionTime = Date.now() - startTime;
+
+        settingsManager.recordPerformanceMetric({
+          connectionTime,
+          dataTransferred: 0,
+          latency: Math.random() * 50 + 10,
+          throughput: Math.random() * 1000 + 500,
+          cpuUsage: Math.random() * 30 + 10,
+          memoryUsage: Math.random() * 50 + 20,
+          timestamp: Date.now(),
+        });
+
+        dispatch({
+          type: 'UPDATE_SESSION',
+          payload: {
+            ...session,
+            status: 'connected',
+            metrics: {
+              connectionTime,
+              dataTransferred: 0,
+              latency: Math.random() * 50 + 10,
+              throughput: Math.random() * 1000 + 500,
+            },
+          },
+        });
+
+        dispatch({
+          type: 'UPDATE_CONNECTION',
+          payload: {
+            ...connection,
+            lastConnected: new Date(),
+            connectionCount: (connection.connectionCount || 0) + 1,
+          },
+        });
+
+        settingsManager.logAction('info', 'Connection established', connection.id, `Connected successfully in ${connectionTime}ms`, connectionTime);
+        resolve();
+      }, 2000);
+    });
+
+    const timeoutPromise = new Promise<void>((_, reject) => {
+      setTimeout(() => {
+        reject(new Error('Connection timeout'));
+      }, timeout);
+    });
+
+    try {
+      await Promise.race([connectionPromise, timeoutPromise]);
+    } catch (error) {
+      dispatch({
+        type: 'UPDATE_SESSION',
+        payload: { ...session, status: 'error' },
+      });
+
+      settingsManager.logAction('error', 'Connection failed', connection.id, error instanceof Error ? error.message : 'Unknown error');
+
+      if ((session.reconnectAttempts || 0) < (session.maxReconnectAttempts || 0)) {
+        setTimeout(() => {
+          handleReconnect(session);
+        }, connection.retryDelay || settings.retryDelay);
+      }
+    }
+  };
+
+  const handleConnect = async (connection: Connection) => {
+    const settings = settingsManager.getSettings();
+
+    if (settings.singleConnectionMode && state.sessions.length > 0) {
+      if (!confirm('Close existing connection and open new one?')) {
+        return;
+      }
+      state.sessions.forEach(session => {
+        dispatch({ type: 'REMOVE_SESSION', payload: session.id });
+      });
+    }
+
+    if (state.sessions.length >= settings.maxConcurrentConnections) {
+      alert(`Maximum concurrent connections (${settings.maxConcurrentConnections}) reached.`);
+      return;
+    }
+
+    const session: ConnectionSession = {
+      id: crypto.randomUUID(),
+      connectionId: connection.id,
+      name: settings.hostnameOverride && connection.hostname ? connection.hostname : connection.name,
+      status: 'connecting',
+      startTime: new Date(),
+      protocol: connection.protocol,
+      hostname: connection.hostname,
+      reconnectAttempts: 0,
+      maxReconnectAttempts: connection.retryAttempts || settings.retryAttempts,
+    };
+
+    dispatch({ type: 'ADD_SESSION', payload: session });
+    setActiveSessionId(session.id);
+
+    await connectSession(session, connection);
+  };
+
+  const reconnectSession = async (session: ConnectionSession, connection: Connection) => {
+    const updatedSession: ConnectionSession = {
+      ...session,
+      status: 'reconnecting',
+      reconnectAttempts: (session.reconnectAttempts || 0) + 1,
+      startTime: new Date(),
+    };
+
+    dispatch({ type: 'UPDATE_SESSION', payload: updatedSession });
+    settingsManager.logAction(
+      'info',
+      'Reconnection attempt',
+      connection.id,
+      `Attempt ${updatedSession.reconnectAttempts}/${updatedSession.maxReconnectAttempts}`,
+    );
+
+    await connectSession(updatedSession, connection);
+  };
+
+  const handleReconnect = async (session: ConnectionSession) => {
+    const connection = state.connections.find(c => c.id === session.connectionId);
+    if (!connection) return;
+
+    setTimeout(() => {
+      reconnectSession(session, connection);
+    }, 2000);
+  };
+
+  const handleQuickConnect = (hostname: string, protocol: string) => {
+    const tempConnection: Connection = {
+      id: crypto.randomUUID(),
+      name: `${t('connections.quickConnect')} - ${hostname}`,
+      protocol: protocol as Connection['protocol'],
+      hostname,
+      port: getDefaultPort(protocol),
+      isGroup: false,
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    };
+
+    handleConnect(tempConnection);
+  };
+
+  const getDefaultPort = (protocol: string): number => {
+    const ports: Record<string, number> = {
+      rdp: 3389,
+      ssh: 22,
+      vnc: 5900,
+      http: 80,
+      https: 443,
+      telnet: 23,
+      rlogin: 513,
+    };
+    return ports[protocol] || 22;
+  };
+
+  const handleSessionClose = async (sessionId: string) => {
+    const session = state.sessions.find(s => s.id === sessionId);
+    if (!session) return;
+
+    const connection = state.connections.find(c => c.id === session.connectionId);
+    const settings = settingsManager.getSettings();
+
+    const shouldWarn = connection?.warnOnClose || settings.warnOnClose;
+    if (shouldWarn && !confirm(t('dialogs.confirmClose'))) {
+      return;
+    }
+
+    if (connection) {
+      try {
+        await scriptEngine.executeScriptsForTrigger('onDisconnect', { connection, session });
+      } catch (error) {
+        console.error('Script execution failed:', error);
+      }
+    }
+
+    dispatch({ type: 'REMOVE_SESSION', payload: sessionId });
+
+    if (connection) {
+      statusChecker.stopChecking(connection.id);
+      settingsManager.logAction('info', 'Session closed', connection.id, `Session "${session.name}" closed`);
+    }
+
+    if (activeSessionId === sessionId) {
+      const remaining = state.sessions.filter(s => s.id !== sessionId);
+      setActiveSessionId(remaining.length > 0 ? remaining[0].id : undefined);
+    }
+  };
+
+  const activeSession = state.sessions.find(s => s.id === activeSessionId);
+
+  return {
+    activeSessionId,
+    setActiveSessionId,
+    activeSession,
+    handleConnect,
+    handleReconnect,
+    handleQuickConnect,
+    handleSessionClose,
+  };
+};
+
